64 bits ids implemented. Optional use FIFO for closing trades implmented. Return type for CreateMarketOrder()  and CreatMarketOrderIfTouched() changed.

+++ Rabun.Oanda.Rest/Models/PositionClosed.cs
@@ -4,7 +4,7 @@ namespace Rabun.Oanda.Rest.Models
 {
     public class PositionClosed
     {
-        public int[] Ids { get; set; }
+        public long[] Ids { get; set; }
         public string Instrument { get; set; }
         public int TotalUnits { get; set; }
         public float Price { get; set; }
