+++ Rabun.Oanda.Rest/Models/TransactionMarginCloseOut.cs
@@ -1,9 +1,10 @@
 ﻿namespace Rabun.Oanda.Rest.Models
 {
-public class TransactionMarginCloseOut: TransactionSimple {
-    public int TradeId;
-    public float Pl;
-    public int Interest;
-    public float AccountBalance;
-}
+    public class TransactionMarginCloseOut : TransactionSimple
+    {
+        public long TradeId;
+        public float Pl;
+        public int Interest;
+        public float AccountBalance;
+    }
 }
