+++ Rabun.Oanda.Rest/Models/TransactionStopLossField.cs
@@ -2,7 +2,7 @@
 {
     public class TransactionStopLossField : TransactionSimple
     {
-        public int TradeId { get; set; }
+        public long TradeId { get; set; }
         public float Pl { get; set; }
         public int Interest { get; set; }
         public float AccountBalance { get; set; }
